16_Langchain_AI_Agents_05-01-2026/langchainaiagent/pyproject.toml
[project]
name = "05-langchain-24-11-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.10.0",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-openai>=1.1.0",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
]
